requirements.txt
# ══════════════════════════════════════════════════════════════════════════
# requirements.txt
# ══════════════════════════════════════════════════════════════════════════
# Framework
Django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.*

# ASGI / Channels / WebSocket
daphne==4.*
channels==4.*
channels-redis==4.*

# Celery / Beat
celery==5.*
django-celery-beat==2.*
django-celery-results==2.*

# Cache / Redis
django-redis==5.*
redis==5.*

# HTTP client
httpx==0.27.*

# SQLite: no extra driver needed — built into Python's stdlib.
# For WAL-mode thread safety with Celery, install this optional package:
# pysqlite3-binary==0.5.*   # only needed if your Python < 3.12 ships an old SQLite

# Auth
djangorestframework-simplejwt==5.*

# Firebase (push notifications)
firebase-admin==6.*

# Utilities
python-decouple==3.*
whitenoise==6.*
Brotli==1.*

# i18n
django-rosetta==0.10.*

# CORS
django-cors-headers==4.*

# Dev extras
# django-debug-toolbar==4.*
